pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["worksheetai"]
package-dir = {"" = "src"}
include-package-data = true

[project]
name = "worksheetai"
version = "0.1.0"
description = "AI-powered worksheet generator"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "llama-index>=0.12.16",
    "openai>=1.3.0",
    "python-dotenv>=0.21.0",
    "questionary>=2.0.0",
]

[project.scripts]
worksheetai = "worksheetai.cli:main"

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]
